+++ mayan/apps/documents/apps.py
@@ -92,10 +92,11 @@ from .links.document_type_links import (
     link_document_type_setup
 )
 from .links.document_version_links import (
-    link_document_version_cache_purge, link_document_version_create,
-    link_document_version_delete, link_document_version_edit,
-    link_document_version_export, link_document_version_list,
-    link_document_version_multiple_delete, link_document_version_return_list,
+    link_document_version_active, link_document_version_cache_purge,
+    link_document_version_create, link_document_version_delete,
+    link_document_version_edit, link_document_version_export,
+    link_document_version_list, link_document_version_multiple_delete,
+    link_document_version_return_list,
     link_document_version_return_to_document, link_document_version_preview,
     link_document_version_print_form,
     link_document_version_transformations_clear,
@@ -600,7 +601,12 @@ class DocumentsApp(MayanAppConfig):
             ), label=_('Pages'), source=DocumentVersion
         )
         SourceColumn(
-            attribute='comment', is_sortable=True, source=DocumentVersion
+            attribute='active', include_label=True, is_sortable=True,
+            source=DocumentVersion
+        )
+        SourceColumn(
+            attribute='comment', include_label=True, is_sortable=True,
+            source=DocumentVersion
         )
 
         # DocumentVersionPage
@@ -853,6 +859,7 @@ class DocumentsApp(MayanAppConfig):
         )
         menu_object.bind_links(
             links=(
+                link_document_version_active,
                 link_document_version_cache_purge,
                 link_document_version_delete, link_document_version_edit,
                 link_document_version_export,