+++ mayan/apps/documents/views/document_version_views.py
@@ -1,5 +1,6 @@
 import logging
 
+from django.conf import settings
 from django.contrib import messages
 from django.template import RequestContext
 from django.urls import reverse
@@ -11,8 +12,8 @@ from mayan.apps.converter.permissions import (
 )
 from mayan.apps.file_caching.tasks import task_cache_partition_purge
 from mayan.apps.views.generics import (
-    FormView, MultipleObjectConfirmActionView, MultipleObjectDeleteView,
-    SingleObjectCreateView, SingleObjectDetailView,
+    ConfirmView, FormView, MultipleObjectConfirmActionView,
+    MultipleObjectDeleteView, SingleObjectCreateView, SingleObjectDetailView,
     SingleObjectEditView, SingleObjectListView
 )
 from mayan.apps.views.mixins import ExternalObjectMixin
@@ -77,6 +78,28 @@ class DocumentVersionCachePartitionPurgeView(MultipleObjectConfirmActionView):
                     #'user_id': self.request.user.pk
                 }
             )
+
+
+class DocumentVersionActiveView(ExternalObjectMixin, ConfirmView):
+    external_object_class = DocumentVersion
+    external_object_permission = permission_document_version_edit
+    external_object_pk_url_kwarg = 'document_version_id'
+
+    def get_extra_context(self):
+        return {
+            'object': self.external_object,
+            'title': _(
+                'Make the document version "%s" the active version?'
+            ) % self.external_object,
+        }
+
+    def view_action(self, form=None):
+        self.external_object.active_set()
+        messages.success(
+            message=_(
+                'Successfully changed the active document version.'
+            ), request=self.request
+        )
 
 
 class DocumentVersionCreateView(ExternalObjectMixin, SingleObjectCreateView):