+++ mayan/apps/ocr/tests/test_models.py
@@ -22,7 +22,7 @@ class GermanOCRSupportTestCase(GenericDocumentTestCase):
     test_document_path = TEST_DEU_DOCUMENT_PATH
 
     def test_ocr_language_backends_end(self):
-        content = self.test_document.latest_version.pages.first().ocr_content.content
+        content = self.test_document.version_active.pages.first().ocr_content.content
 
         self.assertTrue(
             TEST_DOCUMENT_VERSION_OCR_CONTENT_DEU_1 in content